+++ testsuite/src/Issue17.hs
@@ -15,12 +15,12 @@ import           Test.HUnit.Base        (assertBool, assertEqual)
 #if MIN_VERSION_stm(2,5,0)
 main :: IO ()
 main = do
-  -- New queue capacity is set to a negative numer
-  queueIO <- newTBQueueIO (-1 :: Int)
+  -- New queue capacity is set to 0
+  queueIO <- newTBQueueIO 0
   assertNoCapacityTBQueue queueIO
 
-  -- Same as above, except created within STM and different negative number
-  queueSTM <- atomically $ newTBQueue (minBound :: Int)
+  -- Same as above, except created within STM
+  queueSTM <- atomically $ newTBQueue 0
   assertNoCapacityTBQueue queueSTM
 
 assertNoCapacityTBQueue :: TBQueue Int -> IO ()