+++ changelog.md
@@ -6,6 +6,11 @@
 
   * Add `lengthTBQueue` to `Control.Concurrent.STM.TBQueue`.
 
+  * Switched `newTBQueue` and `newTBQueueIO` to accept `Natural` as
+    size. [#17](https://github.com/haskell/stm/pull/17)
+
+  * Dropped support for `base < 4.8`, due to above change.
+
 ----
 
 #### 2.4.5.1 *Sep 2018*